# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAI API Service
  version: 0.1.0
  description: |
    Vercel-first TypeScript gateway for OpenAI Responses API and Audio transcription APIs.

    ## OpenAI API references
    - Responses API: https://platform.openai.com/docs/api-reference/responses
    - Audio transcriptions: https://platform.openai.com/docs/api-reference/audio/createTranscription
    - Models: https://platform.openai.com/docs/models
servers:
  - url: https://your-service.vercel.app
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    LlmRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
          description: "OpenAI model ID, e.g. gpt-5.4 or gpt-5-mini-2025-08-07"
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
        instructions:
          type: string
        reasoning:
          type: object
        tools:
          type: array
          items:
            type: object
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_output_tokens:
          type: integer
          minimum: 1
          maximum: 128000
        metadata:
          type: object
          additionalProperties:
            type: string
        stream:
          type: boolean
          description: "If true, response is sent as Server-Sent Events."
        text:
          type: object
        tool_choice:
          oneOf:
            - type: string
            - type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
              format: uuid
    UsageItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        endpoint:
          type: string
        model:
          type: string
        http_status:
          type: integer
        duration_ms:
          type: integer
        input_tokens:
          type: integer
          nullable: true
        output_tokens:
          type: integer
          nullable: true
        total_tokens:
          type: integer
          nullable: true
        total_cost_usd:
          type: number
          nullable: true
    UsageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UsageItem'
        limit:
          type: integer
        offset:
          type: integer
    AdminUsageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/UsageItem'
              - type: object
                properties:
                  client_id:
                    type: string
                    format: uuid
        limit:
          type: integer
        offset:
          type: integer
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string
                  requestId:
                    type: string
                  now:
                    type: string
                    format: date-time
  /v1/llm:
    post:
      summary: Proxy an OpenAI Responses API request
      description: |
        Forwards the request body to OpenAI's Responses API (`POST /v1/responses`).
        When `stream: true`, the response is returned as Server-Sent Events (SSE).
        The gateway records the full request/response, token usage, and estimated cost.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmRequest'
      responses:
        '200':
          description: OpenAI response payload (JSON or SSE stream)
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Model not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/whisper:
    post:
      summary: Proxy an OpenAI transcription request
      description: |
        Forwards audio to OpenAI's Audio API (`POST /v1/audio/transcriptions`).
        Accepts file upload or a remote URL. SSRF-protected URL fetching is built in.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [model]
              properties:
                model:
                  type: string
                  description: "e.g. whisper-1, gpt-4o-transcribe, gpt-4o-mini-transcribe"
                file:
                  type: string
                  format: binary
                audio_url:
                  type: string
                  format: uri
                  description: "Remote URL to fetch audio from (alternative to file upload)"
      responses:
        '200':
          description: OpenAI transcription payload
        '400':
          description: Validation or audio error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Audio file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/usage:
    get:
      summary: List usage for the authenticated client
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Usage entries for the calling client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/usage:
    get:
      summary: List usage across all clients (admin only)
      description: Requires the SERVICE_ADMIN_KEY as the bearer token.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Usage entries across all clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsageResponse'
        '403':
          description: Forbidden — invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'